extra validation on sku to check whether it matches a sku in database. database re-exported with no allow null for  and allows for longer

+++ Controllers/ProductController.js
@@ -11,7 +11,6 @@ let ProductController = {
         res.json(JsonResService(true, 'successfully retrieved all product data', 200, result))
     },
 
-
     addProduct: async (req, res) => {
         let productToAdd = {
             productName: req.body.productName,
@@ -34,8 +33,13 @@ let ProductController = {
         let sku = req.body.sku
         if (validateProduct.validateSku(sku)) {
             let connection = await dbConnection()
-            await ProductService.deleteProduct(connection, sku)
-            res.json(JsonResService(true, 'successfully deleted product in database', 200, []))
+            let result = await connection.query("SELECT `sku` FROM `products` WHERE `sku` = '" + sku + "';")
+            if (result.length) {
+                await ProductService.deleteProduct(connection, sku)
+                res.json(JsonResService(true, 'successfully deleted product in database', 200, []))
+            } else {
+                res.json(JsonResService(false,  'error: SKU not found in database - no product deleted in database', 404, []))
+            }
         } else {
             res.json(JsonResService(false,  'error: invalid SKU - no product deleted in database', 400, []))
         }
